materialize-sql: do conditional rollbacks on panic
If a panic occurs, any conditional rollbacks must be performed and it is
unlikely that err will be set.

This is a similar fix to 5d9d092.
