+++ materialize-sqlserver/driver.go
@@ -468,31 +468,6 @@ func (d *transactor) Load(it *m.LoadIterator, loaded func(int, json.RawMessage) 
 	return nil
 }
 
-// TxDefuser.MaybeRollback calls Rollback on a Tx, unless Defuse has previously
-// been called.
-type TxDefuser struct {
-	defused bool
-	txn     *stdsql.Tx
-}
-
-func NewTxDefuser(txn *stdsql.Tx) *TxDefuser {
-	return &TxDefuser{
-		defused: false,
-		txn:     txn,
-	}
-}
-
-func (t *TxDefuser) MaybeRollback() {
-	if t.defused {
-		return
-	}
-	t.txn.Rollback()
-}
-
-func (t *TxDefuser) Defuse() {
-	t.defused = true
-}
-
 func (d *transactor) Store(it *m.StoreIterator) (m.StartCommitFunc, error) {
 	ctx := it.Context()
 	txn, err := d.store.conn.BeginTx(ctx, &stdsql.TxOptions{})
@@ -502,7 +477,7 @@ func (d *transactor) Store(it *m.StoreIterator) (m.StartCommitFunc, error) {
 
 	// If a Tx is open and you attempt to close the connection, msqldb will
 	// deadlock.  This handles the case of both an error or a panic occurring.
-	defuser := NewTxDefuser(txn)
+	defuser := sql.NewTxDefuser(txn)
 	defer defuser.MaybeRollback()
 
 	// The mssql driver uses prepared statements to drive bulk inserts. A bulk
